+++ packages/giu-examples/pages/datatable.js
@@ -23,7 +23,6 @@ import {
 } from 'giu';
 import faker from 'faker';
 import 'giu/lib/css/giu.css';
-import 'giu/lib/css/colorInput.css';
 
 faker.seed(0);