+++ packages/giu-examples/pages/material.js
@@ -46,7 +46,6 @@ import HintExample from '../components/demo1-hints';
 import FormExample from '../components/demo1-forms';
 import FormExample2 from '../components/demo1-forms2';
 import 'giu/lib/css/giu.css';
-import 'giu/lib/css/colorInput.css';
 
 // Protect with try .. catch, since MDL assumes that it lives in the
 // browser and crashes during SSR